requirements.txt
pytest
llama-index-llms-ollama